refactor: extract artifact manifest contracts
